Improve controllers behavior

+++ controllers/vultrmachine_controller.go
@@ -125,14 +125,16 @@ func (r *VultrMachineReconciler) Reconcile(req ctrl.Request) (_ ctrl.Result, ret
 func (r *VultrMachineReconciler) reconcileDelete(machineScope *scope.MachineScope) (ctrl.Result, error) {
 	log.Info("Reconciling Machine Delete")
 
-	server, err := r.findServer(machineScope.VultrClient, machineScope.VultrCluster, machineScope.VultrMachine)
+	server, err := r.findServer(machineScope)
 	if err != nil {
 		return ctrl.Result{}, err
 	}
 
-	err = machineScope.VultrClient.DeleteServer(server.ID)
-	if err != nil {
-		return ctrl.Result{}, err
+	if server != nil {
+		err = machineScope.VultrClient.DeleteServer(server.ID)
+		if err != nil {
+			return ctrl.Result{}, err
+		}
 	}
 
 	machineScope.VultrMachine.Finalizers = util.Filter(machineScope.VultrMachine.Finalizers, infrav1alpha2.MachineFinalizer)
@@ -169,10 +171,10 @@ func (r *VultrMachineReconciler) reconcileNormal(machineScope *scope.MachineScop
 	return ctrl.Result{}, nil
 }
 
-func (r *VultrMachineReconciler) findServer(vultrClient *vultr.Client, vultrCluster *infrav1alpha2.VultrCluster, vultrMachine *infrav1alpha2.VultrMachine) (*vultr.Server, error) {
+func (r *VultrMachineReconciler) findServer(machineScope *scope.MachineScope) (*vultr.Server, error) {
 	providerID := ""
-	if vultrMachine.Spec.ProviderID != nil {
-		providerID = *vultrMachine.Spec.ProviderID
+	if machineScope.VultrMachine.Spec.ProviderID != nil {
+		providerID = *machineScope.VultrMachine.Spec.ProviderID
 	}
 
 	// Parse the ProviderID.
@@ -183,7 +185,7 @@ func (r *VultrMachineReconciler) findServer(vultrClient *vultr.Client, vultrClus
 
 	// If the ProviderID populated, get the server using the ID.
 	if err == nil {
-		server, err := vultrClient.GetServer(pid.ID())
+		server, err := machineScope.VultrClient.GetServer(pid.ID())
 		if err != nil && err.Error() == "Invalid server." {
 			return nil, nil
 		}
@@ -194,14 +196,14 @@ func (r *VultrMachineReconciler) findServer(vultrClient *vultr.Client, vultrClus
 	}
 
 	// If the ProviderID is empty, try to get the server using tag and name (label).
-	tag := fmt.Sprintf("%s:owned", vultrCluster.Name)
-	servers, err := vultrClient.GetServersByTag(tag)
+	tag := fmt.Sprintf("%s:owned", machineScope.VultrCluster.Name)
+	servers, err := machineScope.VultrClient.GetServersByTag(tag)
 	if err != nil {
 		return nil, err
 	}
 
 	for _, s := range servers {
-		if s.Name == vultrMachine.GetName() {
+		if s.Name == machineScope.VultrMachine.GetName() {
 			return &s, nil
 		}
 	}
@@ -210,7 +212,7 @@ func (r *VultrMachineReconciler) findServer(vultrClient *vultr.Client, vultrClus
 }
 
 func (r *VultrMachineReconciler) getOrCreate(machineScope *scope.MachineScope) (*vultr.Server, error) {
-	server, err := r.findServer(machineScope.VultrClient, machineScope.VultrCluster, machineScope.VultrMachine)
+	server, err := r.findServer(machineScope)
 	if err != nil {
 		return nil, err
 	}
@@ -228,12 +230,19 @@ func (r *VultrMachineReconciler) getOrCreate(machineScope *scope.MachineScope) (
 		}
 
 		options := &vultr.ServerOptions{
-			ReservedIP: machineScope.VultrCluster.Status.APIEndpoints[0].Host,
-			UserData:   string(userdata),
-			SSHKey:     sshKeyID,
-			Tag:        fmt.Sprintf("%s:owned", machineScope.VultrCluster.Name),
+			Hostname: machineScope.Machine.Name,
+			UserData: string(userdata),
+			SSHKey:   sshKeyID,
+			Tag:      fmt.Sprintf("%s:owned", machineScope.VultrCluster.Name),
 		}
 
+		// Set ReservedIP if the Machine has control-plane label
+		labels := machineScope.Machine.GetLabels()
+		if labels["cluster.x-k8s.io/control-plane"] == "true" {
+			options.ReservedIP = machineScope.VultrCluster.Status.APIEndpoints[0].Host
+		}
+
+		// Set ScriptID if the Machine has Vultr Script ID
 		if machineScope.VultrMachine.Spec.ScriptID != 0 {
 			options.Script = machineScope.VultrMachine.Spec.ScriptID
 		}
@@ -241,6 +250,9 @@ func (r *VultrMachineReconciler) getOrCreate(machineScope *scope.MachineScope) (
 		srv, err := machineScope.VultrClient.CreateServer(machineScope.Machine.Name,
 			machineScope.VultrCluster.Spec.Region, machineScope.VultrMachine.Spec.PlanID,
 			machineScope.VultrMachine.Spec.OSID, options)
+		if err != nil {
+			return nil, err
+		}
 
 		server = &srv
 	}
